resolve changes requested
