+++ frontend-opal/src/reducers/root.js
@@ -1,9 +1,14 @@
 import { combineReducers } from 'redux';
 
-import watchLists from './watchlists';
-import searchResult from './searchResult';
+import watchListsPairs from './watchlists';
+import searchResultPairs from './searchResult';
+
+function setReducer(state, action, pairs) {
+  if (action.type in pairs) return pairs[action.type](state, action);
+  return state;
+}
 
 export default combineReducers({
-  watchLists,
-  searchResult
+  watchLists: (state = [], action) => setReducer(state, action, watchListsPairs),
+  searchResult: (state = {}, action) => setReducer(state, action, searchResultPairs)
 });
